[refactor] UI models: group models
