+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/model/consumption/ConsumptionPresentationStyle.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.ui.model
+package com.rwmobi.kunigami.ui.model.consumption
 
 import androidx.compose.runtime.Immutable
 import com.rwmobi.kunigami.domain.model.consumption.ConsumptionDataGroup
